# Begin work on a simple benchmarking framework using perf