--- a/src/nti/externalization/tests/benchmarks/__init__.py
+++ b/src/nti/externalization/tests/benchmarks/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+A package for benchmarking code.
+
+benchmarks should be performed using the `perf module
+<https://perf.readthedocs.io/en/latest/>`_.
+
+The overall design is taken from `performance
+<https://github.com/python/performance/tree/master/performance>`_.
+
+"""
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function