--- a/src/nti/externalization/tests/benchmarks/configure.zcml
+++ b/src/nti/externalization/tests/benchmarks/configure.zcml
@@ -1,0 +1,16 @@
+<configure xmlns="http://namespaces.zope.org/zope"
+		   xmlns:ext="http://nextthought.com/ntp/ext">
+
+	<include package="zope.component" />
+
+	<include package="nti.externalization" file="meta.zcml" />
+	<include package="nti.externalization" />
+
+	<ext:classObjectFactory factory=".objects.SimplestPossibleObject" />
+
+	<ext:registerAutoPackageIO
+		root_interfaces=".interfaces.IRootInterface"
+		modules=".objects"
+		/>
+
+</configure>