**Also** ensure that more externalized objects get a MimeType by default. (This was necessary to be able to benchmark the different registry approaches.)

Refs #36 

Current benchmark output:

* CPython 2.7
```
.....................
bm_simple_iface: toExternalObject: Mean +- std dev: 82.7 us +- 2.1 us
.....................
bm_simple_iface: fromExternalObject: Mean +- std dev: 145 us +- 5 us
.....................
bm_simple_registered_class: toExternalObject: Mean +- std dev: 69.6 us +- 3.3 us
.....................
bm_simple_registered_class: fromExternalObject: Mean +- std dev: 30.5 us +- 1.2 us
```

* CPython 3.6
```
.....................
bm_simple_iface: toExternalObject: Mean +- std dev: 61.3 us +- 2.9 us
.....................
bm_simple_iface: fromExternalObject: Mean +- std dev: 139 us +- 6 us
.....................
bm_simple_registered_class: toExternalObject: Mean +- std dev: 48.7 us +- 1.9 us
.....................
bm_simple_registered_class: fromExternalObject: Mean +- std dev: 47.5 us +- 1.6 us
```

* PyPy 5.8.0
```
.........
bm_simple_iface: toExternalObject: Mean +- std dev: 7.67 us +- 0.69 us
.........
bm_simple_iface: fromExternalObject: Mean +- std dev: 18.8 us +- 0.9 us
.........
bm_simple_registered_class: toExternalObject: Mean +- std dev: 5.94 us +- 0.28 us
........
bm_simple_registered_class: fromExternalObject: Mean +- std dev: 1.22 us +- 0.07 us
```