@jamadden 2017-09-29
There's really not very much of a framework there yet, just two benchmarks and a tiny amount of supporting code. This feels like one of those things that will grow as needed.

@papachoco and @jzuech3 Could you take a look at the change to externalization.py regarding MIMETYPE please?