@jamadden 2017-09-29
> we do the following check
> if CLASS in result and MIMETYPE not in result:
> in a nti.mimetype decorator

Thanks for the pointer to where that is.

In practice, it's going to do exactly the same thing that *this* code does, and `getattr(obj, 'mimeType', None) or `getattr(obj, 'mime_type', None)` in the majority of the cases (anything registered with `ext:registerAutoPackageIO modules=...` will have a mimeType or mime_type set on the class---it's extremely rare to override that on an instance, I think, to set it to empty strings, especially since you'd have to override both), so the results will be the same. So this might save us a function call. 

In fact I wonder how often the latter part of that function (`nti_mimetype_from_object` which deals with `IContentTypeMarker`) actually gets used from the decorator. I'm guessing "not a lot."